+++ src/kaocha/type/ns.clj
@@ -8,16 +8,19 @@
             [kaocha.type :as type]))
 
 (defn ->testable [ns-name]
-  {:kaocha.testable/type :kaocha.type/ns
-   :kaocha.testable/id   (keyword (str ns-name))
-   :kaocha.testable/desc (str ns-name)
-   :kaocha.ns/name       ns-name})
+  {:kaocha.testable/type            :kaocha.type/ns
+   :kaocha.testable/id              (keyword (str ns-name))
+   :kaocha.testable/desc            (str ns-name)
+   :kaocha.testable/parallelizable? true
+   :kaocha.ns/name                  ns-name})
 
 (defn run-tests [testable test-plan fixture-fn]
   ;; It's not guaranteed the the fixture-fn returns the result of calling the
   ;; tests function, so we need to put it in a box for reference.
-  (let [result (atom (:kaocha.test-plan/tests testable))]
-    (fixture-fn #(swap! result testable/run-testables test-plan))
+  (let [result (promise)]
+    (fixture-fn
+     (fn []
+       (deliver result (testable/run-testables-parent testable test-plan))))
     @result))
 
 (defmethod testable/-load :kaocha.type/ns [testable]