+++ test/unit/kaocha/type/ns_test.clj
@@ -66,27 +66,38 @@
                                          :kaocha.result/fail    0}]}
                 (:result
                  (with-test-ctx {:fail-fast? true}
-                   (testable/run testable testable))))))
+                   (testable/run testable testable)))))))
 
-  (let [testable (testable/load {:kaocha.testable/type :kaocha.type/ns
-                                 :kaocha.testable/id   :baz.qux-test
-                                 :kaocha.testable/desc "baz.qux-test"
-                                 :kaocha.ns/name       'baz.qux-test})]
+(deftest run-test-parallel
+  (classpath/add-classpath "fixtures/f-tests")
+
+  (let [testable (testable/load {:kaocha.testable/type    :kaocha.type/clojure.test
+                                 :kaocha.testable/id      :unit
+                                 :kaocha/ns-patterns      ["-test$"]
+                                 :kaocha/source-paths     ["src"]
+                                 :kaocha/test-paths       ["fixtures/f-tests"]
+                                 :kaocha.filter/skip-meta [:kaocha/skip]})]
+
     (is (match? {:kaocha.testable/type :kaocha.type/ns
-                 :kaocha.testable/id   :baz.qux-test
-                 :kaocha.ns/name       'baz.qux-test
+                 :kaocha.testable/id   :foo.bar-test
+                 :kaocha.ns/name       'foo.bar-test
                  :kaocha.ns/ns         ns?
                  :kaocha.result/tests  [{:kaocha.testable/type  :kaocha.type/var
-                                         :kaocha.testable/id    :baz.qux-test/nested-test
-                                         :kaocha.testable/desc  "nested-test"
-                                         :kaocha.var/name       'baz.qux-test/nested-test
+                                         :kaocha.testable/id    :foo.bar-test/a-test
+                                         :kaocha.testable/desc  "a-test"
+                                         :kaocha.var/name       'foo.bar-test/a-test
                                          :kaocha.var/var        var?
                                          :kaocha.var/test       fn?
                                          :kaocha.result/count   1
                                          :kaocha.result/pass    1
-                                         :kaocha.result/error   1
+                                         :kaocha.result/error   0
                                          :kaocha.result/pending 0
                                          :kaocha.result/fail    0}]}
                 (:result
-                  (with-test-ctx {}
-                    (testable/run testable testable)))))))
+                 (with-test-ctx {:fail-fast? true}
+                   (testable/run testable testable)))))
+    (is (not (nil? (:result
+                    (binding [testable/*config* (assoc testable/*config* :parallel true)]
+                      (with-test-ctx {:fail-fast? true
+                                      :parallel true}
+                        (testable/run testable testable)))))))))